recognition-service/requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
python-dotenv
mysql-connector-python
requests
Pillow
insightface

